FIXED: Compiler warnings by Ulrich Arnold
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1451 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/base/PdfParser.cpp
@@ -1242,7 +1242,7 @@ void PdfParser::CheckEOFMarker()
     const size_t nEOFTokenLen = 5;
     char pszBuff[nEOFTokenLen+1];
 
-    m_device.Device()->Seek(-nEOFTokenLen, std::ios_base::end );
+    m_device.Device()->Seek(-static_cast<int>(nEOFTokenLen), std::ios_base::end );
     if( IsStrictParsing() )
     {
         // For strict mode EOF marker must be at the very end of the file
